Feedback al usuario :mega:
Correccion admin Django con herramienta :bug:
Favicon agregado
Canbio de estilo recomendada por auditoria de colores
modal al elimininar herramienta
estado historico agregado a modelo herramienta

+++ posts/views.py
@@ -13,15 +13,14 @@ from django.contrib import messages
 from posts.models import Perfil, Herramienta
 from decouple import config
 
-# Create your views here.
 
 def index(request):
     herramientas = Herramienta.objects.all()
-
     context = {'herramientas': herramientas}
     return render(request, 'index.html',context)
 
-#autenticación
+
+# Autenticación
 def login_view(request):
     if request.user.is_authenticated():
         return redirect(reverse('catalogo:index'))
@@ -36,27 +35,30 @@ def login_view(request):
         else:
             mensaje = 'Nombre de usuario o clave no valido'
     return render(request, 'login.html', {'mensaje': mensaje})
+
+
 def logout_view(request):
     logout(request)
     return HttpResponseRedirect(reverse('catalogo:index'))
 
-#Herramientas
+
+# Herramientas
 def herramienta_create(request):
     if request.method == 'POST':
         form = HerramientaForm(request.POST, request.FILES)
         if form.is_valid():
             cleaned_data = form.cleaned_data
             nombre = cleaned_data.get('nombre')
-            sistemaOperativo = cleaned_data.get('sistemaOperativo')
+            sistemaoperativo = cleaned_data.get('sistemaOperativo')
             plataforma = cleaned_data.get('plataforma')
-            fichaTecnica = cleaned_data.get('fichaTecnica')
+            fichatecnica = cleaned_data.get('fichaTecnica')
             licencia = cleaned_data.get('licencia')
             estado = 1
             revisor1 = 0
             revisor2 = 0
             autor = request.user.id
             descripcion = cleaned_data.get('descripcion')
-            urlReferencia = cleaned_data.get('urlReferencia')
+            urlreferencia = cleaned_data.get('urlReferencia')
             logo = 'default'
             id_anterior = 0
 
@@ -72,15 +74,20 @@ def herramienta_create(request):
                 url = guardarDarUrl(file, myfile.name)
                 logo = url
 
-            herramienta = Herramienta.objects.create(id_anterior=id_anterior,nombre=nombre,
-                                                     sistemaOperativo=sistemaOperativo,plataforma=plataforma,fichaTecnica=fichaTecnica,licencia=licencia,estado=estado,
-                                                     revisor1=revisor1, revisor2=revisor2, autor=autor, descripcion=descripcion,urlReferencia=urlReferencia,logo=logo)
+            herramienta = Herramienta.objects.create(id_anterior=id_anterior, nombre=nombre,
+                                                     sistemaOperativo=sistemaoperativo, plataforma=plataforma,
+                                                     fichaTecnica=fichatecnica, licencia=licencia, estado=estado,
+                                                     revisor1=revisor1, revisor2=revisor2, autor=autor,
+                                                     descripcion=descripcion, urlReferencia=urlreferencia, logo=logo)
             herramienta.save()
-
-            return HttpResponseRedirect(reverse('catalogo:index'))
+            messages.success(request, 'Se ha creado con éxito la herramienta ' +
+                             herramienta.nombre + ', los cambios serán publicados hasta terminar el proceso de vigía',
+                             extra_tags='alert alert-success')
+            return redirect(reverse('catalogo:index'))
     else:
         form = HerramientaForm()
     return render(request, 'herramienta_create.html', {'form': form})
+
 
 def herramienta_update(request, pk):
     herramienta = Herramienta.objects.get(id=pk)
@@ -90,7 +97,6 @@ def herramienta_update(request, pk):
         if form.is_valid():
             herramienta.estado = 4
             herramienta.save()
-
             cleaned_data = form.cleaned_data
             nombre = cleaned_data.get('nombre')
             sistemaOperativo = cleaned_data.get('sistemaOperativo')
@@ -116,35 +122,87 @@ def herramienta_update(request, pk):
             revisor2 = 0
             estado = 1
             autor = request.user.id
-
-            herramienta_new = Herramienta.objects.create(id_anterior=id_anterior, nombre=nombre,
-                                                     sistemaOperativo=sistemaOperativo, plataforma=plataforma,
-                                                     fichaTecnica=fichaTecnica, licencia=licencia, estado=estado,
-                                                     revisor1=revisor1, revisor2=revisor2, autor=autor,
-                                                     descripcion=descripcion, urlReferencia=urlReferencia, logo=logo)
-            herramienta_new.save()
-
+            herramienta_n = Herramienta.objects.create(id_anterior=id_anterior, nombre=nombre,
+                                                       sistemaOperativo=sistemaOperativo, plataforma=plataforma,
+                                                       fichaTecnica=fichaTecnica, licencia=licencia, estado=estado,
+                                                       revisor1=revisor1, revisor2=revisor2, autor=autor,
+                                                       descripcion=descripcion, urlReferencia=urlReferencia, logo=logo)
+            herramienta_n.save()
+            messages.success(request, 'Se ha editado con éxito la herramienta '+
+                             herramienta.nombre+', los cambios serán publicados hasta terminar el proceso de vigía',
+                             extra_tags='alert alert-success')
             return render(request, 'herramienta_detail.html',{'herramienta': herramienta})
     else:
         form = HerramientaUpdateForm(instance=herramienta)
     return render(request, 'herramienta_update.html', {'form': form, 'id':pk})
 
-def herramienta_delete(request, pk):
-    herramienta = Herramienta.objects.get(id=pk)
-    if request.method == 'POST':
-        herramienta.delete()
-        return HttpResponseRedirect(reverse('catalogo:index'))
-    return render(request, 'herramienta_delete.html',{'object': herramienta})
 
 def herramienta_detail(request, pk):
     herramienta = Herramienta.objects.get(id=pk)
-    context = {'herramienta': herramienta}
-    return render(request, 'herramienta_detail.html', context)
-
-
-def herramientas_estados(request):
     if request.user.is_authenticated():
-        herramientas_r = Herramienta.objects.all().filter(estado=1).exclude(autor=request.user.id).exclude(revisor1=request.user.id)
+        if request.method == 'POST' and request.user.perfil.role == 1:
+            herramienta.delete()
+            messages.success(request, 'Ha eliminado con éxito a ' + herramienta.nombre,
+                             extra_tags='alert alert-success')
+            return redirect(reverse('catalogo:index'))
+        elif herramienta.estado == 3:
+            context = {'herramienta': herramienta}
+            return render(request, 'herramienta_detail.html', context)
+
+        elif herramienta.estado == 4:
+            temp = list(Herramienta.objects.all().filter(id_anterior=pk).exclude(estado=5))
+            herramienta_old = temp.pop()
+            autor = Perfil.objects.get(id=herramienta_old.autor)
+            estado = 'en revisión' if herramienta_old.estado == 1 else 'pendiente a publicar se'
+            msg = 'Esta es una versión que será modificada, ' + autor.user.first_name +\
+                ' ' + autor.user.last_name + ' la ha modificado y está ' + estado +\
+                  '. por tal motivo no es posible crear ediciones'
+            context = {'herramienta': herramienta, 'herramienta_old': herramienta_old}
+            messages.warning(request,msg, extra_tags='alert alert-warning')
+            return render(request, 'herramienta_detail.html', context)
+
+        elif herramienta.estado == 1 or herramienta.estado == 2:
+            if herramienta.id_anterior == 0:
+                context = {'herramienta': herramienta}
+                return render(request, 'herramienta_detail.html', context)
+            else:
+                try:
+                    herramienta_old = Herramienta.objects.get(id=herramienta.id_anterior, estado=4)
+                    autor = Perfil.objects.get(id=herramienta.autor)
+                    context = {'herramienta': herramienta, 'herramienta_old': herramienta_old}
+                    msg = 'Esta modificación fue realizada por: ' + autor.user.first_name + ' ' + autor.user.last_name
+
+                    if not herramienta.revisor1 == 0 and not herramienta.revisor2 == 0:
+                        revisor1 = Perfil.objects.get(id=herramienta.revisor1)
+                        revisor2 = Perfil.objects.get(id=herramienta.revisor2)
+                        msg = msg + ' y los revisores  fueron: ' + revisor1.user.first_name + ' ' +\
+                                    revisor1.user.last_name + ' y ' + revisor2.user.first_name \
+                                    + ' ' + revisor2.user.last_name
+
+                    messages.warning(request, msg, extra_tags='alert alert-warning')
+                    return render(request, 'herramienta_detail.html', context)
+                except Herramienta.DoesNotExist:
+                    print("Base de datos inconsistente")
+                except Herramienta.MultipleObjectsReturned:
+                    print("Base de datos inconsistente")
+        else:
+            if herramienta.estado == 3:
+                context = {'herramienta': herramienta}
+                return render(request, 'herramienta_detail.html', context)
+
+    else:
+        if herramienta.estado == 3 or herramienta.estado == 4:
+            context = {'herramienta': herramienta}
+            return render(request, 'herramienta_detail.html', context)
+
+
+
+
+
+def herramientas_vigia(request):
+    if request.user.is_authenticated():
+        herramientas_r = Herramienta.objects.all().filter(estado=1).exclude(autor=request.user.id). \
+            exclude(revisor1=request.user.id)
         herramientas_p = Herramienta.objects.all().filter(estado=2)
         context = {'herramientas_r': herramientas_r, 'herramientas_p': herramientas_p}
         return render(request, 'vigia.html', context)
@@ -153,7 +211,8 @@ def herramientas_estados(request):
         context = {'herramientas': herramientas}
         return render(request, 'index.html', context)
 
-def herramienta_revisar(request,pk):
+
+def herramienta_revisar(request, pk):
     if request.user.is_authenticated():
         herramienta = Herramienta.objects.get(id=pk)
         if herramienta.revisor1 == 0:
@@ -162,7 +221,7 @@ def herramienta_revisar(request,pk):
             herramienta.revisor2 = request.user.id
             herramienta.estado = 2
         herramienta.save()
-        messages.success(request, 'Revisado con exito', extra_tags='alert alert-success')
+        messages.success(request, 'Ha revisado con éxito a '+herramienta.nombre, extra_tags='alert alert-success')
         return redirect(reverse('catalogo:vigia'))
 
     else:
@@ -170,13 +229,13 @@ def herramienta_revisar(request,pk):
         context = {'herramientas': herramientas}
         return render(request, 'index.html', context)
 
+
 def herramienta_publicar(request,pk):
     if request.user.is_authenticated():
         herramienta = Herramienta.objects.get(id=pk)
         herramienta.estado = 3
-        messages.success(request, 'Publicado con exito', extra_tags='alert alert-success')
+        messages.success(request, 'Ha sido publicado con éxito a '+herramienta.nombre, extra_tags='alert alert-success')
 
-        print("herramienta.id_anterior" + str(herramienta.id_anterior))
         if not herramienta.id_anterior == 0:
             herramienta_delete = Herramienta.objects.get(id=herramienta.id_anterior)
             herramienta_delete.estado = 5
@@ -258,7 +317,6 @@ def user_update(request, pk):
                 user_model.username = username
 
             user_model.save()
-            print(user_model.username + " " + username)
 
             profile_model = Perfil.objects.get(user_id=user_model.id)
             fotoSubio = True if 'foto' in request.FILES else False
@@ -298,7 +356,6 @@ def user_updateGTI(request, pk):
                 user_model.username = username
 
             user_model.save()
-            print(user_model.username + " " + username)
 
             profile_model = Perfil.objects.get(user_id=user_model.id)
             fotoSubio = True if 'foto' in request.FILES else False
